[0.0.3] Adding the punch modules 'modules' that lists all available modules.
